// apps/backend/src/generated/prisma/enums.ts
export const ExamPartType = {
  WRITTEN: 'WRITTEN',
  PRACTICAL: 'PRACTICAL',
} as const;

export type ExamPartType = (typeof ExamPartType)[keyof typeof ExamPartType];

export const QuestionType = {
  SINGLE_CHOICE: 'SINGLE_CHOICE',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  SHORT_ANSWER: 'SHORT_ANSWER',
} as const;

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];
// ...
export const AttemptStatus = {
  NOT_STARTED: 'NOT_STARTED',
  NOT_ATTENDED: 'NOT_ATTENDED',
  IN_PROGRESS: 'IN_PROGRESS',
  SUBMITTED: 'SUBMITTED',
  INCOMPLETE: 'INCOMPLETE',
  GRADING: 'GRADING',
  GRADED: 'GRADED',
} as const;

export type AttemptStatus = (typeof AttemptStatus)[keyof typeof AttemptStatus];

export const SubmissionMethod = {
  MANUAL: 'MANUAL',
  AUTO_TIMEOUT: 'AUTO_TIMEOUT',
  AUTO_PART_CLOSED: 'AUTO_PART_CLOSED',
} as const;

export type SubmissionMethod =
  (typeof SubmissionMethod)[keyof typeof SubmissionMethod];

export const PassStatus = {
  PENDING: 'PENDING',
  PASS: 'PASS',
  FAIL: 'FAIL',
} as const;

export type PassStatus = (typeof PassStatus)[keyof typeof PassStatus];